pyproject.toml
[project]
name = "eccv"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers",
    "huggingface-hub>=0.36.0",
    "numpy>=2.2.6",
    "opencv-python>=4.13.0.90",
    "peft>=0.18.1",
    "pillow>=12.1.0",
    "safetensors>=0.7.0",
    "sentencepiece>=0.2.1",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers==4.47.1",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers.git" }
transformers = { git = "https://github.com/huggingface/transformers.git" }
